requirements.txt
Flask==2.2.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.0
bcrypt==3.2.0
PyJWT==2.1.0